+++ deployment/docker_build_branch.sh
@@ -19,23 +19,19 @@ export PRG_PATH=$(pwd)
 
 SVC_NAME="jobs"
 SVC1="jobsapi"
-SVC2="jobsmigrate"
 SVC3="jobslib"
 SVC3TAG="jobsworker"
 REPO="tapis"
 BRANCH_NAME="$1"
 
 export SVC1_WAR=v3
-export SVC2_JAR=${SVC2}.jar
 export SVC3_JAR=shaded-${SVC3}.jar
 
 # Directories for docker build context. Dockerfiles are located here in repo
 BUILD_DIR1=$PRG_PATH/tapis-${SVC1}
-BUILD_DIR2=$PRG_PATH/tapis-${SVC2}
 BUILD_DIR3=$PRG_PATH/tapis-${SVC3TAG}
 # Target build directories containing compiled code and jar/war files
 BUILD_DIRT1=$PRG_PATH/../tapis-${SVC1}/target
-BUILD_DIRT2=$PRG_PATH/../tapis-${SVC2}/target
 BUILD_DIRT3=$PRG_PATH/../tapis-${SVC3}/target
 ENV=$1
 
@@ -56,10 +52,6 @@ if [ ! -d "$BUILD_DIRT1" ]; then
   echo "Build directory missing. Please build. Directory: $BUILD_DIRT1"
   exit 1
 fi
-if [ ! -d "$BUILD_DIRT2" ]; then
-  echo "Build directory missing. Please build. Directory: $BUILD_DIRT2"
-  exit 1
-fi
 if [ ! -d "$BUILD_DIRT3" ]; then
   echo "Build directory missing. Please build. Directory: $BUILD_DIRT3"
   exit 1
@@ -71,11 +63,9 @@ GIT_BRANCH_LBL=$(awk '{print $1}' $BUILD_DIRT1/classes/git.info)
 GIT_COMMIT_LBL=$(awk '{print $2}' $BUILD_DIRT1/classes/git.info)
 
 TAG_BRANCH1="${REPO}/${SVC1}:${BRANCH_NAME}"
-TAG_BRANCH2="${REPO}/${SVC2}:${BRANCH_NAME}"
 TAG_BRANCH3="${REPO}/${SVC3TAG}:${BRANCH_NAME}"
 
 export BUILD_TIME1="$(awk '{print $1}' ${BUILD_DIRT1}/classes/build.time)"
-export BUILD_TIME2="$(awk '{print $1}' ${BUILD_DIRT2}/classes/build.time)"
 export BUILD_TIME3="$(awk '{print $1}' ${BUILD_DIRT3}/classes/build.time)"
 
 # If branch name is UNKNOWN or empty as might be the case in a jenkins job then
@@ -116,25 +106,6 @@ docker build -f Dockerfile \
    --build-arg BUILD_TIME=$BUILD_TIME -t "${TAG_BRANCH1}" .
 
 
-# Build image for jobsmigrate
-echo "======================================================================"
-echo "Building local image for service ${SVC2} using tag: $TAG_BRANCH2"
-echo "======================================================================"
-# Move to the build directory
-cd $BUILD_DIR2 || exit
-echo "**** Removing any old service jar files from Docker build context"
-rm -f $SVC2_JAR
-echo "**** Copying $SVC2_JAR to Dockerfile context"
-cp $BUILD_DIRT2/${SVC2_JAR} $BUILD_DIR2
-
-echo "**** Building docker image: $TAG_BRANCH2"
-docker build -f Dockerfile \
-   --build-arg SRVC_JAR=${SVC2_JAR} --build-arg VER=$VER --build-arg GIT_COMMIT=$GIT_COMMIT \
-   --build-arg BUILD_TIME=$BUILD_TIME -t "${TAG_BRANCH2}" .
-
-echo "**** Removing jar from Docker build context"
-rm ${BUILD_DIR2}/${SVC2_JAR}
-
 # Build image for jobsworker
 echo "======================================================================"
 echo "Building local image for service ${SVC3TAG} using primary tag: $TAG_BRANCH3"
@@ -159,7 +130,6 @@ if [ "x$2" = "x-push" ]; then
   echo "Pushing images to docker hub."
   # NOTE: Use current login. Jenkins job does login
   docker push "$TAG_BRANCH1"
-  docker push "$TAG_BRANCH2"
   docker push "$TAG_BRANCH3"
 fi