+++ tapis-jobslib/src/main/java/edu/utexas/tacc/tapis/jobs/events/JobEventManager.java
@@ -3,7 +3,6 @@ package edu.utexas.tacc.tapis.jobs.events;
 import java.sql.Connection;
 import java.util.List;
 
-import edu.utexas.tacc.tapis.jobs.utils.JobUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,9 +17,11 @@ import edu.utexas.tacc.tapis.jobs.model.enumerations.JobEventType;
 import edu.utexas.tacc.tapis.jobs.model.enumerations.JobStatusType;
 import edu.utexas.tacc.tapis.jobs.queue.JobQueueManager;
 import edu.utexas.tacc.tapis.jobs.queue.JobQueueManagerNames;
+import edu.utexas.tacc.tapis.jobs.utils.JobUtils;
 import edu.utexas.tacc.tapis.shared.exceptions.TapisException;
 import edu.utexas.tacc.tapis.shared.exceptions.runtime.TapisRuntimeException;
 import edu.utexas.tacc.tapis.shared.i18n.MsgUtils;
+import edu.utexas.tacc.tapis.sharedapi.security.ResourceRequestUser;
 
 /** This class records noteworthy job events and asynchronously send notifications
  * to subscribers.  The threads that call a record method synchronously write to the
@@ -112,7 +113,7 @@ public final class JobEventManager
      * @param conn existing connection or null
      * @throws TapisException on error
      */
-    public JobEvent recordStatusEvent(Job job, JobStatusType newStatus, 
+    public JobEvent recordStatusEvent(ResourceRequestUser rUser, Job job, JobStatusType newStatus,
                                       JobStatusType oldStatus, Connection conn)
      throws TapisException
     {
@@ -132,7 +133,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, conn);
+        _jobEventsDao.createEvent(rUser, jobEvent, conn);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -148,7 +149,6 @@ public final class JobEventManager
      * @param job the job that generated the event
      * @param transferStatus status of transfer as reported by Files
      * @param transactionId transaction id from the Files service
-     * @param conn existing connection or null
      * @throws TapisException on error
      */
     public JobEvent recordStagingInputsEvent(Job job, TransferStatusEnum transferStatus, 
@@ -170,12 +170,11 @@ public final class JobEventManager
                 " in tenant " + job.getTenant() + ".";
         
         // Fill in the event details as a JSON object.
-        var data = JobEventData.getTransferEventData(job, msg, transferStatus,
-                                                     transactionId);
+        var data = JobEventData.getTransferEventData(job, msg, transferStatus, transactionId);
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, null);
+        _jobEventsDao.createEvent(null, jobEvent, null);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -191,7 +190,6 @@ public final class JobEventManager
      * @param job the job that generated the event
      * @param transferStatus status of transfer as reported by Files
      * @param transactionId transaction id from the Files service
-     * @param conn existing connection or null
      * @throws TapisException on error
      */
     public JobEvent recordArchivingEvent(Job job, TransferStatusEnum transferStatus, 
@@ -218,7 +216,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, null);
+        _jobEventsDao.createEvent(null, jobEvent, null);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -248,7 +246,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
 		
 		// Save in db.
-		_jobEventsDao.createEvent(jobEvent, null);
+		_jobEventsDao.createEvent(null, jobEvent, null);
 		postEventToNotificationService(jobEvent);
 		return jobEvent;
     }
@@ -278,7 +276,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
 		// Save in db.
-		_jobEventsDao.createEvent(jobEvent, null);
+		_jobEventsDao.createEvent(null, jobEvent, null);
 		postEventToNotificationService(jobEvent);
 		return jobEvent;
    }
@@ -314,7 +312,7 @@ public final class JobEventManager
        jobEvent.setDescription(data);
        
        // Save in db and send to notifications service asynchronously.
-       _jobEventsDao.createEvent(jobEvent, null);
+       _jobEventsDao.createEvent(null, jobEvent, null);
        postEventToNotificationService(jobEvent);
        return jobEvent;
    }
@@ -370,7 +368,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, conn);
+        _jobEventsDao.createEvent(null, jobEvent, conn);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -386,7 +384,6 @@ public final class JobEventManager
      * @param jobUuid the job that generated the event
      * @param tenant the job tenant
      * @param action added or removed
-     * @param conn existing connection or null
      * @throws TapisException on error
      */
     public JobEvent recordSubscriptionEvent(String jobUuid, String tenant, 
@@ -413,7 +410,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, null);
+        _jobEventsDao.createEvent(null, jobEvent, null);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -449,7 +446,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, null);
+        _jobEventsDao.createEvent(null, jobEvent, null);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -464,7 +461,6 @@ public final class JobEventManager
      * 
      * @param job the job that generated the event
      * @param finalMsg the job's final message
-     * @param conn existing connection or null
      * @throws TapisException on error
      */
     public JobEvent recordFinalMessageEvent(Job job, String finalMsg)
@@ -487,7 +483,7 @@ public final class JobEventManager
         jobEvent.setDescription(data);
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, null);
+        _jobEventsDao.createEvent(null, jobEvent, null);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -522,7 +518,7 @@ public final class JobEventManager
         jobEvent.setEventDetail(eventDetail); 
         
         // Save in db and send to notifications service asynchronously.
-        _jobEventsDao.createEvent(jobEvent, conn);
+        _jobEventsDao.createEvent(null, jobEvent, conn);
         postEventToNotificationService(jobEvent);
         return jobEvent;
     }
@@ -541,7 +537,6 @@ public final class JobEventManager
      * @param sender the tapis user that sent the event
      * @param eventData the user provided body of the event
      * @param eventDetail the event subtype or key
-     * @param conn existing connection or null
      * @return the JobEvent on success or null on failure
      * @throws TapisException on error
      */
@@ -576,15 +571,14 @@ public final class JobEventManager
      */
     private boolean postEventToNotificationService(JobEvent jobEvent)
     {
-        // Error already logged.
-        try {JobQueueManager.getInstance().postEventQueue(jobEvent);}
-            catch (Exception e) {
-            	String msg = JobUtils.getMsg("JOBS_EVENT_POST_ERROR",
-            			        jobEvent.getEvent().name(),  
-            			        JobQueueManagerNames.getEventQueueName());
-            	_log.warn(msg, e);
-            	return false;
-            }
-        return true; // success
+      // Error already logged.
+      try {JobQueueManager.getInstance().postEventQueue(jobEvent);}
+      catch (Exception e) {
+        String msg = JobUtils.getMsg("JOBS_EVENT_POST_ERROR", jobEvent.getEvent().name(),
+                                     JobQueueManagerNames.getEventQueueName());
+         _log.warn(msg, e);
+         return false;
+      }
+      return true; // success
     }
 }