+++ tapis-jobslib/src/test/java/edu/utexas/tacc/tapis/jobs/dao/CreateJobWithConstraintsTest.java
@@ -24,7 +24,7 @@ public class CreateJobWithConstraintsTest
 		// Insert job record.
 		for (int i = 0; i < ITERATIONS; i++) {
 			Job job = initJob(i);
-			dao.createJob(job);
+			dao.createJob(null, job);
 		}
 		
 		// Get all job records.